Data structure Linked List deletion operation Program other cases implemented
Implementation of deletion operation in linked using C language. his Program deletes the node at the head of the linked list. Also implemented the function to delete node at specific index in the linked list. Also implemented the use case of deleting the node at the end of the linked list. Also implemented the use case of deleting the node with a specific value